% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLabel.r
\name{addLabel}
\alias{addLabel}
\title{Query Data ILOSTAT}
\usage{
addLabel(df, Lang = "en", Drop = FALSE)
}
\arguments{
\item{df}{a dataframe comming from getData function.}

\item{Drop}{code, default = FALSE .}

\item{lang}{character define one of the 3 official ILO language en, fr, es.}
}
\description{
Query data from ILOSTAT SDMX API
}
\details{
Helper function to efficiently query data from ILOSTAT SDMX API.
}
\examples{
################################## fetch data and add labels

# example with attribute
res <- getData("DF_YI_AFG_EMP_TEMP_SEX_AGE_NB/....")
res_label <- AddLabel(res)
# example data only
res <- getData("DF_YI_AFG_ALL/.....?detail=dataonly")
res_label <- AddLabel(res)
}
\author{
ILO bescond
}
\seealso{
\code{\link{getData}} \code{\link{getCodelist}} \code{\link{getDataStructure}}
}
\keyword{ILO,}
\keyword{R}
\keyword{SDMX,}

