% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addres_geocode.R
\name{addres_geocode}
\alias{addres_geocode}
\title{Gives the geocode of the address given by the user}
\usage{
addres_geocode(address)
}
\arguments{
\item{address}{character parameter under the format "street number street type street name postcode city name"}
}
\value{
numeric vector
}
\description{
Gives the geocode of the address given by the user
}
\examples{
\dontrun{
assert_geocode(12 rue de Rivoli 75001 Paris)}

}
