% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_info.R
\name{address_info}
\alias{address_info}
\title{Gives the cadaster information of an address}
\usage{
address_info(polygon_ref, feuille_ref, matrix_parcelles, matrix_feuilles)
}
\arguments{
\item{polygon_ref}{numeric}

\item{feuille_ref}{numeric}

\item{matrix_parcelles}{matrix corresponding to the given address (parcelles category) in the cadastre database}

\item{matrix_feuilles}{matrix corresponding to the given address (feuilles category) in the cadastre database}
}
\value{
character string
}
\description{
Gives the cadaster information of an address
}
\examples{
\dontrun{
adress_info(1,1,parcelles75107.json,feuilles75107.json)
}

}
