% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_to_cadastre.R
\name{address_to_cadastre}
\alias{address_to_cadastre}
\title{Gives the map and all the cadaster information of a given address- global function}
\usage{
address_to_cadastre(address)
}
\arguments{
\item{address}{character string}
}
\description{
Gives the map and all the cadaster information of a given address- global function
}
