% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\title{Aligns together either a single drug regimen/record, or a list of regimens against a drug record,
returning either a set of global alignments or a set of local alignments, either overlapping
or otherwise}
\usage{
align(regimen, regName, drugRec, g, Tfac, s = NA, verbose, mem, removeOverlap)
}
\arguments{
\item{regimen}{A sequence to be aligned, usually a short sequence defining a chemotherapy regimen,
may contain multiple regimens if supplied as a list}

\item{regName}{The name of the supplied regimen/s, may be a character or a list}

\item{drugRec}{A longer sequence to be aligned, usually a patient's converted drug occurrence table}

\item{g}{A gap penalty supplied to the temporal needleman wunsch/smith waterman algorithms}

\item{Tfac}{The time penalty factor. All time penalties are calculated as a percentage of Tfac}

\item{s}{A substituion matrix, either user-defined or derived from defaultSmatrix. Will be auto-generated if left blank.}

\item{verbose}{A variable indicating how verbose the python script should be in reporting results
Verbose = 0 : Print nothing
Verbose = 1 : Print seqs and scores
Verbose = 2 : Report seqs, scores, H and traceMat}

\item{mem}{A number defining how many sequences to hold in memory during local alignment.
Mem = -1 : Script will define memory length according to floor(len(regimen)/len(drugRec))
Mem = 0 : Script will return exactly 1 alignment
Mem = 1 : Script will return 1 alignment and all alignments with the same score
Mem = X : Script will return X alignments and all alignments with equivalent score as the Xth alignment}

\item{removeOverlap}{A variable indicating whether to remove overlaps (1) or leave them in the output data (0)}
}
\value{
dat A dataframe containing information on the resulting alignments
}
\description{
Aligns together either a single drug regimen/record, or a list of regimens against a drug record,
returning either a set of global alignments or a set of local alignments, either overlapping
or otherwise
}
\examples{
output <- align(regimen,drugRec)
}
