% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentSummaryReport.R
\name{anyUncoveredDrugs}
\alias{anyUncoveredDrugs}
\title{Any uncovered drugs?}
\usage{
anyUncoveredDrugs(pa)
}
\arguments{
\item{pa}{Processed alignment data with personID and t_start, and one patient.}
}
\description{
This function checks if there are any uncovered drugs for a single patient. 
It is a part of nonRegimenDrugExposure function.
}
