% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Artemis.r
\name{Artemis}
\alias{Artemis}
\title{Rilostat: Query data from ILOSTAT SDMX rest API}
\description{
dplyr provides a flexible grammar of data manipulation. It's the next
iteration of plyr, focused on tools for working with data frames (hence the
\emph{d} in the name).
}
\details{
Targets:

\itemize{
\item Identify and return datastructure, codelist and data and metadata
\item from http://www.ilo.org/ilostat/sdmx/ws/rest/
}
}

