% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{as.data.frame.eDNA_simulation}
\alias{as.data.frame.eDNA_simulation}
\title{Methods for eDNA simulations}
\usage{
\method{as.data.frame}{eDNA_simulation}(x)
}
\arguments{
\item{x}{object of class eDNA_simulation}
}
\value{
data.frame
}
\description{
Methods for eDNA simulations
}
\details{
Methods for eDNA simulations
}
\author{
Matt Espe
}
