% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preAlign.R
\name{buildBlacklistRegex}
\alias{buildBlacklistRegex}
\title{Build a single regex from blacklist terms}
\usage{
buildBlacklistRegex(blacklist_set)
}
\arguments{
\item{blacklist_set}{Character vector of blacklist terms.}
}
\value{
Regex pattern string.
}
\description{
Build a single regex from blacklist terms
}
