% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessAlign.R
\name{calculateRegimenTimes}
\alias{calculateRegimenTimes}
\title{Calculates Regimen start and end time.}
\usage{
calculateRegimenTimes(df, drugDF)
}
\arguments{
\item{df}{A dataframe with drugRec_Start and drugRec_End columns}

\item{drugDF}{A drugDF object generated by createDrugDF()}
}
\value{
A data.table object without removed overlaps
}
\description{
Adds cumulative times from drugDF to the dataframe,
where each regimen starts and ends
}
