% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanDataFrame.r
\name{cleanDataFrame}
\alias{cleanDataFrame}
\title{Query Data ILOSTAT}
\usage{
cleanDataFrame(df, header = TRUE, tolower = FALSE)
}
\arguments{
\item{df, }{data frame to clean.}

\item{header, }{header yet define by default TRUE,  or store at the first row.}
}
\description{
Clean dataframe, return a tbl, with character colums, clean up NA and trim characters
}
\details{
Helper function to efficiently query data from ILOSTAT SDMX API.
}
\author{
ILO bescond
}
\seealso{
\code{\link{getData}} \code{\link{getDataStructure}}
}
\keyword{ILO,}
\keyword{R}
\keyword{SDMX,}

