% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessAlign.R
\name{combineOverlaps}
\alias{combineOverlaps}
\title{Combine Overlaps}
\usage{
combineOverlaps(df, regimenCombine)
}
\arguments{
\item{regimenCombine}{Allowed days between same regimen before being combined}

\item{output}{An output dataframe created by align()}
}
\description{
Merge overlapping regimens in alignment output.
Regimens are only merged if they are identical and occur consecutively in time.
}
