% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_models.R
\name{compile_models}
\alias{compile_models}
\title{Compile models}
\usage{
compile_models(
  model_names = c("eDNA_lm.stan", "eDNA_lmer.stan", "eDNA_sim_omni.stan"),
  models = system.file("stan", model_names, package = "artemis"),
  cache_dir = tools::R_user_dir("artemis", "cache"),
  rewrite = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{model_names}{the names of the models to compile}

\item{models}{full file path to the location of models}

\item{rewrite}{logical, whether to re-compile the models}

\item{...}{additional arguments passed to \code{cmdstanr::cmdstan_model()}}
}
\description{
This compiles the models needed for the package.
}
\details{
Compile models.
}
\author{
Matt Espe
}
