% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_models.R
\name{compiled_models_ok}
\alias{compiled_models_ok}
\title{Check Compiled Models}
\usage{
compiled_models_ok(
  model_names = c("eDNA_lm", "eDNA_lmer", "eDNA_sim_omni"),
  cache_dir = getOption("artemis_cache_dir", R_user_dir("artemis", "cache"))
)
}
\arguments{
\item{model_names}{Character string of the base-names of the
compiled models. On Windows platforms, the extension "exe" is
added}

\item{cache_dir}{directory where the compiled models should be
found.}
}
\value{
logical, TRUE if the compiled models were found and FALSE
    otherwise
}
\description{
Checks if the pre-compiled models for the artemis package exist.
}
\examples{
{
compiled_models_ok()
}
}
\author{
Matt Espe
}
