% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{conf_to_probs}
\alias{conf_to_probs}
\title{Confidence to probability}
\usage{
conf_to_probs(conf)
}
\arguments{
\item{conf}{numeric between 0 and 1}
}
\value{
vector of length 2
}
\description{
Confidence level to probability
}
\details{
This is a convience function to convert from a confidence level to
the lower and upper probability of the confidence interval.
}
\examples{

conf_to_probs(0.95)
}
\author{
Matt Espe
}
