% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cq_to_lnconc}
\alias{cq_to_lnconc}
\title{Convert CQ value to [eDNA]}
\usage{
cq_to_lnconc(Cq_values, std_curve_alpha, std_curve_beta)
}
\arguments{
\item{Cq_values}{numeric vector, value of CQ}

\item{std_curve_alpha}{the alpha (intercept) value for the standard curve}

\item{std_curve_beta}{the beta (slope) value for the standard curve}
}
\value{
numeric vector, [eDNA] values
}
\description{
Convenience function to convert from Cq to [eDNA]
}
\details{
Convenience function to convert from Cq to [eDNA]
}
\author{
Matt Espe
}
