% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessAlign.R
\name{createDrugDF}
\alias{createDrugDF}
\title{Creates cumulative times for each drug in the drug record}
\usage{
createDrugDF(patientDrugRecord)
}
\arguments{
\item{patientDrugRecord}{A complete patient drug record}
}
\value{
A data.table object with drugs
}
\description{
Creates cumulative times for each drug in the drug record
}
