% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{decode}
\alias{decode}
\title{Decodes a tuple of tuples into a TxDrugA.TyDrugB text format}
\usage{
decode(s1)
}
\arguments{
\item{s1}{A list of tuples, each tuple containing a time and drug specification}
}
\value{
str_seq - Encoded string sequence
s_encoded <- encode(s1)
}
\description{
Decodes a tuple of tuples into a TxDrugA.TyDrugB text format
}
