% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preAlign.R
\docType{data}
\name{default_concept_list}
\alias{default_concept_list}
\title{Map concept IDs to concept names
Default concept IDs when no file is provided.}
\format{
An object of class \code{numeric} of length 43.
}
\usage{
default_concept_list
}
\description{
Map concept IDs to concept names
Default concept IDs when no file is provided.
}
\keyword{datasets}
