% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultSmatrix.R
\name{defaultSmatrix}
\alias{defaultSmatrix}
\title{Create a default substitution matrix from all unique elements of s1 and s2}
\usage{
defaultSmatrix(s1, s2)
}
\arguments{
\item{s1}{Sequence 1, typically a HemOnc regimen}

\item{s2}{Sequence 2, typically an encoded set of drug occurrences for a patient}
}
\value{
s - A dataframe object filled with a score of -1.1 in all non-diagonal cells and a score of 1 on the diagonal.
}
\description{
Create a default substitution matrix from all unique elements of s1 and s2
}
\examples{
s <- defaultSmatrix(s1, s2)
}
