% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_json}
\alias{df_json}
\title{A default cohort indicating metastatic lung cancer patients}
\format{
A data frame with 1 rows and 4 variables:
}
\usage{
df_json
}
\description{
A default cohort data.frame with json
}
\keyword{datasets}
