% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputOutput.R
\name{dfFromCDM}
\alias{dfFromCDM}
\title{Generate a con_df dataframe using CDMConnector}
\usage{
dfFromCDM(cdm, cohortName, cdmSchema = NULL, writeSchema = NULL)
}
\arguments{
\item{cdm}{An OMOP cdm reference}

\item{cohortName}{Name of cohort table}

\item{cdmSchema}{A schema containing a valid OMOP CDM}

\item{writeSchema}{A schema where the user has write access}
}
\value{
A con_df dataframe
}
\description{
Generate a con_df dataframe using CDMConnector
}
