% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{eDNA_model-class}
\alias{eDNA_model-class}
\title{eDNA model fit results}
\description{
Placeholder.
}
\section{Slots}{

\describe{
\item{\code{ln_conc}}{matrix, the estimated latent variable, log(eDNA concentration)}

\item{\code{Cq_star}}{matrix, the predicted CQ value for each obs}

\item{\code{betas}}{matrix, the posterior estimate for each beta in the model}

\item{\code{sigma_Cq}}{the estimated measurement error on CQ}

\item{\code{formula}}{the formula used to fit the model}

\item{\code{x}}{the model.matrix used to fit the model}

\item{\code{std_curve_alpha}}{the alpha for the std. curve conversion formual used}

\item{\code{std_curve_beta}}{the beta for the std. curve conversion formula used}

\item{\code{upper_Cq}}{the upper limit for CQ}

\item{\code{stanfit}}{the result from rstan::sampling}
}}

