% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{eDNA_model_lm-class}
\alias{eDNA_model_lm-class}
\alias{eDNA_model_lmer-class}
\title{eDNA simulation results}
\description{
Placeholder

Placeholder
}
\section{Slots}{

\describe{
\item{\code{random_x}}{data.frame of the grouping variables used}

\item{\code{random_sd}}{the estimated stdev. of each of the random effects}
}}

