% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{eDNA_simulation_lmer-class}
\alias{eDNA_simulation_lmer-class}
\alias{eDNA_simulation_lm-class}
\title{eDNA simulation results}
\description{
Placeholder

Placeholder
}
\section{Slots}{

\describe{
\item{\code{groups}}{the grouping variables used}

\item{\code{random_sd}}{the stdev of the random effects}
}}

