% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{encode}
\alias{encode}
\title{Encodes a sequence from txDRUG.tyDRUG format into a tuple of tuples}
\usage{
encode(str_seq)
}
\arguments{
\item{str_seq}{String sequence to decode}
}
\value{
s1 - A list of tuples, each tuple containing a time and drug specification
}
\description{
Encodes a sequence from txDRUG.tyDRUG format into a tuple of tuples
}
\examples{
s_decoded <- decode(str_seq)
}
