% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preAlign.R
\name{escapeRegex}
\alias{escapeRegex}
\title{Escape regex metacharacters}
\usage{
escapeRegex(text)
}
\arguments{
\item{text}{Character vector to escape.}
}
\value{
Character vector with regex metacharacters escaped.
}
\description{
Escape regex metacharacters
}
