% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputOutput.R
\name{filter_stringDF}
\alias{filter_stringDF}
\title{Filter a stringDF dataframe to contain only valid patients}
\usage{
filter_stringDF(stringDF, min)
}
\arguments{
\item{stringDF}{A stringDF dataframe}

\item{min}{Number of valid drugs in a given subject's drug record required to pass
the filter}
}
\value{
A filtered dataframe containing the relevant patients and their
drug exposure strings
}
\description{
Filter a stringDF dataframe to contain only valid patients
}
