% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{generateRegimenStats}
\alias{generateRegimenStats}
\title{Generate Regimen Stats}
\usage{
generateRegimenStats(pa)
}
\arguments{
\item{pa}{A dataframe processed alignments, generated by processAlignments()}
}
\value{
A data frame containing summary stats from processed regimen data
importFrom dplyr group_by summarise full_join count pivot_wider
}
\description{
Generates a data frame containing summary stats from processed regimen data.
}
