% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCodelist.r
\name{getCodelist}
\alias{getCodelist}
\title{Query Data ILOSTAT}
\usage{
getCodelist(codelist, lang = "en", name = FALSE, test = "-test")
}
\arguments{
\item{codelist}{a codelist as returned by \code{scheme="datastructure", dimension="codelist"}.}

\item{lang}{return label in es, fr or en .}

\item{name}{return just codelist name default FALSE.}

\item{test}{internal ILO test parameter.}
}
\description{
Query data from ILOSTAT SDMX API
}
\details{
Helper function to efficiently query data from ILOSTAT SDMX API.
}
\examples{
################################# use to identify correct 'DSD'		 

# fetch collection define on ILOSTAT
res <- getCodelist(codelist = "CL_COLLECTION", lang ="en")	
					
# fetch country available on ILOSTAT
res <- getCodelist(codelist = "CL_COUNTRY", lang ="es")							

# fetch indicator define in YI collection (code, Label in french), also available en (English), es (Spanish)
res <- getCodelist(codelist = "CL_INDICATOR_YI", lang ="fr") 
}
\author{
ILO bescond
}
\seealso{
\code{\link{getData}} \code{\link{getDataStructure}}
}
\keyword{ILO,}
\keyword{R}
\keyword{SDMX,}

