% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputOutput.R
\name{getCohortSet}
\alias{getCohortSet}
\title{Generate a conDF dataframe from a valid CDM connection and
a valid cohort JSON.}
\usage{
getCohortSet(cdm, json, name)
}
\arguments{
\item{cdm}{A valid CDM connection object}

\item{json}{A valid Cohort specification JSON}

\item{name}{A name for the given cohort}
}
\value{
A cdm dataframe object
}
\description{
Generate a conDF dataframe from a valid CDM connection and
a valid cohort JSON.
}
