% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputOutput.R
\name{getConDF}
\alias{getConDF}
\title{Generate a con_df dataframe without using CDMConnector}
\usage{
getConDF(connectionDetails, json, name, cdmSchema, writeSchema)
}
\arguments{
\item{connectionDetails}{A set of DatabaseConnector connectiondetails}

\item{json}{A loaded cohort from loadJSON()}

\item{name}{A cohort-specific name for written tables}

\item{cdmSchema}{A schema containing a valid OMOP CDM}

\item{writeSchema}{A schema where the user has write access}
}
\value{
A con_df dataframe
}
\description{
Generate a con_df dataframe without using CDMConnector
}
