% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getConditions}
\alias{getConditions}
\title{Print a list of valid conditions for use by getRegimens}
\usage{
getConditions()
}
\value{
condList
}
\description{
Print a list of valid conditions for use by getRegimens
}
