% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCount.r
\name{getCount}
\alias{getCount}
\title{Query Data ILOSTAT}
\usage{
getCount(DSD, test = "-test")
}
\arguments{
\item{DSD}{a datastructure definition identified by the triplet \code{[collection; country; indicator]}. Arguments description come from 'http://www.ilo.org/ilostat/content/conn/ILOSTATContentServer/path/Contribution Folders/statistics/web_pages/static_pages/technical_page/ilostat_appl/SDMX_User_Guide.pdf' .}
}
\description{
Query data from ILOSTAT SDMX API
}
\details{
Helper function to efficiently query data from ILOSTAT SDMX API.
}
\examples{
################################## use to check data available

# example with attribute
res <- getCount("YI_AFG_EMP_TEMP_SEX_AGE_NB/....")

# example without attribute
res <- getCount("YI_AFG_ALL/.....?detail=dataonly")

# example of last N data
res <- getCount("YI_AFG_EMP_TEMP_SEX_AGE_NB/.....?lastNObservations=1")   

# example of first N data
res <- getCount("YI_AFG_EMP_TEMP_SEX_AGE_NB/.....?firstNObservations=2")

# example with multi country
res <- getCount("YI_ALL_EMP_TEMP_SEX_AGE_NB/.MEX+ESP")

# check availability of time series	
res <- getCount("YI_ALL_EMP_TEMP_SEX_AGE_NB/.....?detail=serieskeysonly")
### as from 2009
res <- getCount("YI_ALL_EMP_TEMP_SEX_AGE_NB/.....?startPeriod=2009-01-01&detail=serieskeysonly")
}
\author{
ILO bescond
}
\seealso{
\code{\link{getCodelist}} \code{\link{getDataStructure}} \code{\link{getData}}
}
\keyword{ILO,}
\keyword{R}
\keyword{SDMX,}

