% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.r
\name{getData}
\alias{getData}
\title{Query Data ILOSTAT}
\usage{
getData(DSD, test = "-test")
}
\arguments{
\item{DSD}{a datastructure definition identified by the triplet \code{[collection; country; indicator]}. Arguments description come from 'http://www.ilo.org/ilostat/content/conn/ILOSTATContentServer/path/Contribution Folders/statistics/web_pages/static_pages/technical_page/ilostat_appl/SDMX_User_Guide.pdf' .}

\item{test}{internal ILO test parameter.}
}
\description{
Query data from ILOSTAT SDMX API
}
\details{
Helper function to efficiently query data from ILOSTAT SDMX API.
}
\examples{
################################## use to fetch data and metadata related

# example with attribute
res <- getData("YI_AFG_EMP_TEMP_SEX_AGE_NB/....")

# example without attribute
res <- getData("YI_AFG_ALL/......?detail=dataonly")

# example of last N data
res <- getData("YI_AFG_EMP_TEMP_SEX_AGE_NB/......?lastNObservations=1")   

# example of first N data
res <- getData("YI_AFG_EMP_TEMP_SEX_AGE_NB/......?firstNObservations=2")

# example with multi country
res <- getData("YI_ALL_EMP_TEMP_SEX_AGE_NB/.MEX+ESP.....")

# check availability of time series	
res <- getData("YI_FRA_EMP_TEMP_SEX_AGE_NB/......?detail=serieskeysonly")
### as from 2009
res <- getData("YI_ALL_EMP_TEMP_SEX_AGE_NB/......?startPeriod=2009-01-01&detail=serieskeysonly")
}
\author{
ILO bescond
}
\seealso{
\code{\link{getCodelist}} \code{\link{getDataStructure}}
}
\keyword{ILO,}
\keyword{R}
\keyword{SDMX,}

