% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataStructure.r
\name{getDataStructure}
\alias{getDataStructure}
\title{Query Data ILOSTAT}
\usage{
getDataStructure(DSD = "YI_ALB_EAP_TEAP_SEX_AGE_NB",
  dimension = "conceptRef", component = FALSE, lang = "en", test = NULL)
}
\arguments{
\item{DSD}{a datastructure definition identified by the triplet \code{[collection; country; indicator]}.}

\item{dimension}{for scheme "datastructure": the format of dimensions returned. The dimension format "codelist" is compatible with scheme "codelist".}

\item{component}{default FALSE return just the name else all components.}

\item{lang}{if component = TRUE return label in es, fr or en .}

\item{test}{internal ILO test parameter.}
}
\description{
Query data from ILOSTAT SDMX API
}
\details{
Helper function to efficiently query data from ILOSTAT SDMX API.
}
\examples{
################################# use to identify 'codelist'		

# fetch structure of codelists in YI collection 
res <- getDataStructure(dimension = "codelist", DSD = "YI_ALL_ALL") 	

# fetch structure of codelists in YI collection for MEX country
res <- getDataStructure(dimension = "codelist", DSD = "YI_MEX_ALL") 	

# fetch structure of codelists in YI collection for a specific indicator 
res <- getDataStructure(dimension = "codelist", DSD = "YI_ALL_UNE_TUNE_SEX_AGE_NB") 

################################# use to define correct 'filter'

# fetch structure of conceptRef in YI collection 
res <- getDataStructure(DSD = "YI_ALL_ALL", dimension = "conceptRef") 	

res <- getDataStructure(DSD = "YI_ALB_ALL", dimension = "conceptRef")

res <- getDataStructure(DSD = "YI_ALL_EMP_TEMP_SEX_AGE_NB", dimension = "conceptRef")

################################# use to get the all data structure 'filter'

res <- getDataStructure(DSD = "YI_FRA_EMP_TEMP_SEX_AGE_NB", component = TRUE)
names(res) 
## return a list with : 
## --KeyFamily = df definition of this DSD
res$KeyFamily
## --dataStructure = df data structure definition with dimensions and attributes	
res$dataStructure
## --codelist = list of codelists df		
names(res$codelist) 	; res$codelist$CL_COUNTRY	
}
\author{
ILO bescond
}
\seealso{
\code{\link{getData}} \code{\link{getCodelist}}
}
\keyword{ILO,}
\keyword{R}
\keyword{SDMX,}

