% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getPatientData}
\alias{getPatientData}
\title{Create a dataframe containing patient IDs and regimen sequences from a valid CDM connection}
\usage{
getPatientData(cancerID, cdmSchema, dbiconn)
}
\arguments{
\item{cancerID}{The highest-level concept ID which relates to the specific cancer of choice, default is 443392, which includes
all neoplastic diseases.}

\item{cdmSchema}{The name of the cdm schema where sequences will be synthesised from}

\item{dbiconn}{A dbConnect object which contains the user's database connection details
see: https://www.rdocumentation.org/packages/DBI/versions/0.5-1/topics/dbConnect}
}
\value{
seqDF - A dataframe object filled with all relevant patient IDs and a regimen sequence for each patient in a valid format
}
\description{
Create a dataframe containing patient IDs and regimen sequences from a valid CDM connection
}
\examples{
seqDF <- getPatientData(dbiconn, cdmSchema, cancerID)
}
