% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{json}
\alias{json}
\title{A default cohort indicating metastatic lung cancer patients}
\format{
A data frame with 1 rows and 4 variables:
}
\usage{
json
}
\description{
A default cohort json
}
\keyword{datasets}
