% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lnconc_to_cq}
\alias{lnconc_to_cq}
\title{Convert [eDNA] to Cq}
\usage{
lnconc_to_cq(x, std_curve_alpha, std_curve_beta, upper_Cq = 40)
}
\arguments{
\item{x}{[eDNA] values}

\item{std_curve_alpha}{the alpha value for the standard curve}

\item{std_curve_beta}{the beta value for the standard curve}

\item{upper_Cq}{the max Cq value}
}
\value{
vector of Cq values
}
\description{
Convenience function for converting values
}
\details{
Convenience function for converting values
}
\author{
Matt Espe
}
