% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cadastre_sheet.R
\name{load_cadastre_sheet}
\alias{load_cadastre_sheet}
\title{Loads and open the cadaster sheet corresponding to the city code}
\usage{
load_cadastre_sheet(url, city_code)
}
\arguments{
\item{url}{character vecor}

\item{city_code}{numerical vector}
}
\value{
dataframe
}
\description{
Loads and open the cadaster sheet corresponding to the city code
}
\examples{
\dontrun{
load_cadaster_sheet("https://cadastre.data.gouv.fr/data/etalab-cadastre/latest/communes/75/75117/cadastre-75117-batiments.json.gz",75117)
}


}
