% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputOutput.R
\name{loadDrugs}
\alias{loadDrugs}
\title{Generate a set of valid chemotherapy drugs}
\usage{
loadDrugs()
}
\value{
A dataframe of valid drugs and their mappings between HemOnc and OMOP
validDrugs <- validDrugs()
}
\description{
Generate a set of valid chemotherapy drugs
}
