% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputOutput.R
\name{loadRegimens}
\alias{loadRegimens}
\title{Generate a set of regimens, specified by condition}
\usage{
loadRegimens()
}
\value{
A dataframe of HemOnc regimen data, filtered by condition
}
\description{
Generate a set of regimens, specified by condition
}
