% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo.eDNA_model}
\alias{loo.eDNA_model}
\title{Approximate leave-one-out cross-validation}
\usage{
\method{loo}{eDNA_model}(x, ...)
}
\arguments{
\item{x}{a eDNA_model object}

\item{...}{additional args passed to \code{loo}}
}
\value{
a list with the results of the calculation.
}
\description{
A 'loo' method for eDNA_model objects, which is simply a wrapper
around \code{loo(extract_log_lik(x))}
}
\details{
See ?loo::loo for more information about this calculation
}
\author{
Matt Espe
}
