% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_point.R
\name{map_point}
\alias{map_point}
\title{Loads the information of the parcel and displays it on a map of the parcel using its coordinates}
\usage{
map_point(banRmatrix, information)
}
\arguments{
\item{banRmatrix}{banR object}

\item{information}{character vector}
}
\value{
dataframe
}
\description{
Loads the information of the parcel and displays it on a map of the parcel using its coordinates
}
\examples{
\dontrun{
map_point(2.2945, 48.8584, 'Here is the <b>Tour Eiffel</b>, Paris')
}


}
