% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrodataLoad.r
\name{MicrodataLoad}
\alias{MicrodataLoad}
\title{Query Data ILOSTAT}
\usage{
MicrodataLoad(path, resume = FALSE)
}
\arguments{
\item{path, }{path and filenames of the original microdataset.}

\item{resume, }{if true, return list of colnames and label ($Label), colnames and vector class ($Class), colnames and summary for numeric and table for other ($Summary), instead of the full dataset.}
}
\description{
load microdata in various formats .dbf, .dta (13 and before),
}
\details{
Helper function to efficiently load microdataset.
}
\examples{
MicrodataLoad("H:\\\\COMMON\\\\A0 Short term indicators\\\\Collection\\\\COU_ZAF_TEST\\\\Input\\\\qlfs-2009-03.dta", TRUE) \%>\% filter(!value \%in\% c('UQNO', 'Weight')) \%>\% unnest \%>\% ilo::save_ilo(. = 'my_summary')
}
\author{
ILO bescond
}
\seealso{
\code{\link{MicroFrameWork}} \code{\link{MicroProcessTime}}
}
\keyword{ILO,}
\keyword{R}

