% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicroFrameWork.r
\name{MicroFrameWork}
\alias{MicroFrameWork}
\title{Query Data ILOSTAT}
\usage{
MicroFrameWork(Definition, Indicator_Target)
}
\arguments{
\item{Definition, }{dataframe mapping between microdataset and ilo framework.}

\item{Indicator_Target, }{dataframe with ilo indicators target.}
}
\description{
preape the ILO framework, from the national definition and the ilo targeted indicator,
}
\details{
Helper function to efficiently process microdataset.
}
\author{
ILO bescond
}
\seealso{
\code{\link{MicroProcessTime}} \code{\link{MicrodataLoad}}
}
\keyword{ILO,}
\keyword{R}

