% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicroProcessTime.r
\name{MicroProcessTime}
\alias{MicroProcessTime}
\title{Query Data ILOSTAT}
\usage{
MicroProcessTime(File, Frame)
}
\arguments{
\item{File, }{file name(s) of the original microdataset(s), if multiple, ";" separate.}

\item{Frame, }{ilo collection framework from MicroFrameWork function.}
}
\description{
Process one time period from microdataset to ilo framework,
}
\details{
Helper function to efficiently process microdataset.
}
\author{
ILO bescond
}
\seealso{
\code{\link{MicroFrameWork}} \code{\link{MicrodataLoad}}
}
\keyword{ILO,}
\keyword{R}

