% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_polygon.R
\name{nearest_polygon}
\alias{nearest_polygon}
\title{Returns the closes polygon to a long lat point}
\usage{
nearest_polygon(banRmatrix, cadaster_sheet)
}
\arguments{
\item{banRmatrix}{the banR matrix given by address_geocode}

\item{cadaster_sheet}{the corresponding cadaster sheet in dataframe form}
}
\value{
integer
}
\description{
Returns the closes polygon to a long lat point
}
\examples{
\dontrun{
nearest_polygon(geocode_address("5 rue d'Olivet 75007"),cadaster_sheet_75107)
}

}
