% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{no_unique_aligned_drugs}
\alias{no_unique_aligned_drugs}
\title{Calculates the number of unique aligned drugs in a regimen}
\usage{
no_unique_aligned_drugs(regimen)
}
\description{
Calculates the number of unique aligned drugs in a regimen
}
