% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentSummaryReport.R
\name{nonRegimenDrugExposure}
\alias{nonRegimenDrugExposure}
\title{Non-Regimen Drug Exposure?}
\usage{
nonRegimenDrugExposure(pa)
}
\arguments{
\item{pa}{Processed alignment data with personID and t_start.}
}
\description{
This function checks if there are any uncovered drugs for each patient. 
It uses the anyUncoveredDrugs function to check for each patient separately.
}
