% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{normal}
\alias{normal}
\title{Normal prior distribution}
\usage{
normal(location = 0, scale = 1, autoscale = TRUE)
}
\arguments{
\item{location}{numeric, the mean of the normal distribution}

\item{scale}{numeric, the sd of the normal distribution}

\item{autoscale}{logical, whether the priors should be scaled. See
\code{?rstanarm::priors} for details on the scaling.}
}
\value{
named list
}
\description{
Normal prior distribution
}
\details{
Parameters for the normal distribution, to be used for setting
priors on model estimates.
}
\author{
Matt Espe
}
