% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_date}
\alias{normalize_date}
\title{Normalize dates}
\usage{
normalize_date(x, tz = "UTC")
}
\arguments{
\item{x}{Column containing dates}

\item{tz}{Optional. Timezone}
}
\description{
Normalize dates
}
