% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.eDNA_model}
\alias{plot.eDNA_model}
\title{Plot eDNA model results}
\usage{
\method{plot}{eDNA_model}(x, y, pars = "betas", ...)
}
\arguments{
\item{x}{object of class eDNA_model_*}

\item{y}{ignored}

\item{pars}{parameters to plot}

\item{...}{additional args passed to plot.stanfit}
}
\value{
ggplot of posterior estimates
}
\description{
Plot eDNA model results
}
\details{
Plot eDNA model results. Currently, this is just a wrapper for
\code{rstan::plot}, which produces a "catapillar" plot.
}
\author{
Matt Espe
}
