% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.eDNA_p_detect}
\alias{plot.eDNA_p_detect}
\title{Plot p(detect)}
\usage{
\method{plot}{eDNA_p_detect}(x, y, probs = c(0.025, 0.975), ylim = c(0,
  1), point_size = 2.5, n_breaks = 5, ...)
}
\arguments{
\item{x}{an object of class "eDNA_p_detect", produced by \code{est_p_detect}}

\item{y}{ignored}

\item{probs}{numeric vector of length 2, the lower and upper probabilities for
quantiles displayed}

\item{ylim}{the y limits on the plot}

\item{point_size}{point size for the estimate or the mean value}

\item{n_breaks}{passed to \code{pretty()}}

\item{...}{additional args, currently ignored}
}
\value{
a ggplot object
}
\description{
Plot the p(detect)
}
\details{
Plot the p(detect)
}
\author{
Matt Espe
}
