% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.eDNA_simulation}
\alias{plot.eDNA_simulation}
\title{Plot method for eDNA simulations}
\usage{
\method{plot}{eDNA_simulation}(x, y, response = "Cq_star",
  probs = c(0.025, 0.975), ...)
}
\arguments{
\item{x}{object of class eDNA_simulation}

\item{y}{ignored}

\item{response}{the response variable to plot}

\item{probs}{the probability for plotting CIs}

\item{...}{ignored}
}
\value{
a list of ggplots
}
\description{
Plot method for eDNA simulations
}
\details{
Plot method for eDNA simulations
}
\author{
Matt Espe
}
