% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotErasFrequency}
\alias{plotErasFrequency}
\title{Plots a plot displaying the ERA frequency of the top N most frequent eras}
\usage{
plotErasFrequency(processedEras, N = 10)
}
\arguments{
\item{processedEras}{An output dataframe created by calculateEras}

\item{N}{The number of top rows to plot}
}
\description{
Plots a plot displaying the ERA frequency of the top N most frequent eras
}
