% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotOutput}
\alias{plotOutput}
\title{Plots, or returns a plot, displaying a full alignment output}
\usage{
plotOutput(
  output,
  returnPlot = F,
  individual_Tracks = F,
  normScore = F,
  allowOverlaps = F
)
}
\arguments{
\item{output}{An output dataframe created by align()}

\item{returnPlot}{A boolean indicating whether or not to return a ggplot object}

\item{individual_Tracks}{A boolean indicating whether drugs will each have a single track}

\item{normScore}{A boolean indicating whether to use actual scores or scores normalised to
number of aligned drugs}

\item{allowOverlaps}{A boolean indicating whether to leave or remove overlapping regimens}
}
\value{
regPlot - A ggplot object
}
\description{
Plots, or returns a plot, displaying a full alignment output
}
\examples{
plotOutput(output,F,1)
outputPlot <- plotOutput(output,F,1)
}
