% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotProcesssed}
\alias{plotProcesssed}
\title{Plots, or returns a plot, displaying a full alignment output utilising an already
processed output}
\usage{
plotProcesssed(processedAll, fontSize = 2.5)
}
\arguments{
\item{processedAll}{An output dataframe created by processAlignments}

\item{fontSize}{The desired font size of the text}
}
\value{
regPlot - A ggplot object
}
\description{
Plots, or returns a plot, displaying a full alignment output utilising an already
processed output
}
