% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRecord}
\alias{plotRecord}
\title{Plots, or returns a plot, displaying a single drug record}
\usage{
plotRecord(drugRec, returnPlot = F, individual_Tracks = 1)
}
\arguments{
\item{drugRec}{Either an encoded or unencoded regimen sequence}

\item{returnPlot}{A boolean indicating whether or not to return a ggplot object}

\item{individual_Tracks}{A boolean indicating whether drugs will each have a single track}
}
\value{
regPlot - A ggplot object
}
\description{
Plots, or returns a plot, displaying a single drug record
}
\examples{
plotRecord(drugRec,F,1)
drugPlot <- plotRecord(drugRec,F,1)
}
