% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRegimen}
\alias{plotRegimen}
\title{Plots, or returns a plot, displaying a single regimen sequence}
\usage{
plotRegimen(regimen, regimenName, returnPlot = F)
}
\arguments{
\item{regimen}{Either an encoded or unencoded regimen sequence}

\item{regimenName}{The name of the regimen, as specified by the user}

\item{returnPlot}{A boolean indicating whether or not to return a ggplot object}
}
\value{
regPlot - A ggplot object
}
\description{
Plots, or returns a plot, displaying a single regimen sequence
}
\examples{
plotRegimen(regimen,regimenName)
regPlot <- plotRegimen
}
