% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSankey}
\alias{plotSankey}
\title{Plots a sankey diagram displaying the flow between first, second and third regimen eras}
\usage{
plotSankey(processedEras, regGroups)
}
\arguments{
\item{processedEras}{An output dataframe created by calculateEras}

\item{regGroups}{A dataframe indicating how to group regimens}
}
\description{
Plots a sankey diagram displaying the flow between first, second and third regimen eras
}
