% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{postprocessDF}
\alias{postprocessDF}
\title{Postprocess alignment output}
\usage{
postprocessDF(output, regimenCombine = 28)
}
\arguments{
\item{output}{An output dataframe created by align()}

\item{regimenCombine}{Allowed number of days between two instances of the same regimen before}
}
\value{
df - A data.frame object
}
\description{
Postprocess alignment output
}
