% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.eDNA_model}
\alias{predict.eDNA_model}
\alias{predict.eDNA_model_lmer}
\title{Predict eDNA model}
\usage{
\method{predict}{eDNA_model}(
  object,
  newdata = NULL,
  include_sigma = FALSE,
  interval = FALSE,
  interval_fun = posterior_interval,
  ...
)

\method{predict}{eDNA_model_lmer}(object, newdata = NULL, type = c(), ...)
}
\arguments{
\item{object}{an object of class eDNA_lm{er}}

\item{newdata}{optional, data.frame of new observations to predict
values for}

\item{include_sigma}{logical, should the predictions include
measurement error?}

\item{interval}{logical, should the raw predictions be returned
(\code{interval = FALSE}) or should an interval be computed
(\code{interval = TRUE})}

\item{interval_fun}{a function which computes an interval given a
vector of posterior samples}

\item{...}{additional arguments passed to the interval function}

\item{type}{the type of the prediction, MORE HERE}
}
\value{
either a vector of predictions, or a matrix with the
    prediction plus interval, depending on the value of
    \code{interval}
}
\description{
Predict values for eDNA model
}
\details{
Predict methods for an eDNA model fit. Currently, these functions
are quite limited.
}
\author{
Matt Espe
}
