% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.eDNA_model}
\alias{print.eDNA_model}
\title{Print eDNA model results}
\usage{
\method{print}{eDNA_model}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{object of class "eDNA_model_*"}

\item{digits}{number of digits to show}

\item{...}{additional arguments passed to \code{print}}
}
\description{
Print eDNA model results
}
\details{
.. content for \details{} ..
}
\author{
Matt Espe
}
