% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.eDNA_p_detect}
\alias{print.eDNA_p_detect}
\title{Print eDNA p(detect)}
\usage{
\method{print}{eDNA_p_detect}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{object of class "eDNA_p_detect"}

\item{digits}{number of digits to show}

\item{...}{additional arguments passed to \code{print}}
}
\value{
x
}
\description{
Print method for p(detect)
}
\details{
Print method for p(detect)
}
\author{
Matt Espe
}
