% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.eDNA_simulation}
\alias{print.eDNA_simulation}
\title{Print eDNA simulation}
\usage{
\method{eDNA_simulation}{print}(x, FUN = summary,
  digits = getOption("digits"), show_variables = FALSE, ...)
}
\arguments{
\item{x}{object of class eDNA_simulation}

\item{FUN}{a function to use to summarize the results, default is \code{summary}}

\item{digits}{number of digits to show}

\item{show_variables}{logical, should the variable levels used in the simulation
be included}

\item{...}{additional arguments passed to \code{print}}
}
\description{
Print eDNA simulation
}
\details{
Print method.
}
\author{
Matt Espe
}
