% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessAlign.R
\name{processAlignments}
\alias{processAlignments}
\title{Post-process alignment output}
\usage{
processAlignments(ra, regimenCombine, regimens = "none")
}
\arguments{
\item{ra}{Raw alignments. An output dataframe produced by generateRawAlignments()}

\item{regimenCombine}{The numeric value of days allowed between regimens of the same
name before they are collapsed/summarised into a single regimen}

\item{regimens}{The set of input regimens used to generate alignments, from which cycle lengths may be derived}
}
\value{
A dataframe processed alignments
}
\description{
Perform post-processing on a data frame of raw alignment results. Three main steps are performed:
1. Remove overlaps between regimens
2. Combine regimens
3. Calculate additional time-based metrics and line of treatment.
}
