% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ranef,eDNA_model_lmer-method}
\alias{ranef,eDNA_model_lmer-method}
\title{Summarize random effect}
\usage{
\S4method{ranef}{eDNA_model_lmer}(object, FUN = quantile, probs = c(0.025, 0.5, 0.975), ...)
}
\arguments{
\item{object}{an object of class eDNA_model}

\item{FUN}{a posterior summary function, default \code{quantile}}

\item{probs}{probabilities for the posterior summary function}

\item{...}{extra args passed to \code{FUN}}
}
\value{
matrix, with one row per random effect, and one column per
    probability
}
\description{
Summarize random effects
}
\details{
This function returns a summary of the random effects for an
eDNA_model produced by \code{eDNA_lmer}
}
\author{
Matt Espe
}
