% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentSummaryReport.R
\name{regimenEndsBeforeLastDrug}
\alias{regimenEndsBeforeLastDrug}
\title{Regimens ends before last drug?}
\usage{
regimenEndsBeforeLastDrug(pa)
}
\arguments{
\item{pa}{Processed alignment data with personID and t_start}
}
\description{
This function checks if regimens end before the last drug for each patient.
}
