% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regimengroups}
\alias{regimengroups}
\title{A default set of regimen groups, such as chemotherapy, ICI-combinations etc.}
\format{
A data frame with 84 rows and 3 variables:
}
\usage{
regimengroups
}
\description{
A set of default regimen groups
}
\keyword{datasets}
