% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regimens}
\alias{regimens}
\title{Default Regimen data for lung cancer}
\format{
A data frame with 228 rows and 20 variables:
}
\usage{
regimens
}
\description{
A set of default regimens, derived from HemOnc - Thoracic Cancer
}
\keyword{datasets}
