% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentSummaryReport.R
\name{regimenStartAfterFirstDrug}
\alias{regimenStartAfterFirstDrug}
\title{Regimens start after the first drug?}
\usage{
regimenStartAfterFirstDrug(pa)
}
\arguments{
\item{pa}{Processed alignment data with personID and t_start}
}
\description{
This function checks if regimens for each person start after the first drug.
}
