% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentSummaryReport.R
\name{regimenStartBeforeFirstDrug}
\alias{regimenStartBeforeFirstDrug}
\title{Regimens start before the first drug?}
\usage{
regimenStartBeforeFirstDrug(pa, ra)
}
\arguments{
\item{pa}{Processed alignment data with personID and t_start}
}
\description{
This function checks if regimens for each person start after the first drug.
}
