% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{regName}
\alias{regName}
\title{Global Variable Warnings
These warnings serve no purpose and since tidyr 1.2.0 the .data pronoun
is not used in select() and unnest(), thus leading to these warnings
being erroneously generated}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
regName
}
\description{
Global Variable Warnings
These warnings serve no purpose and since tidyr 1.2.0 the .data pronoun
is not used in select() and unnest(), thus leading to these warnings
being erroneously generated
}
\keyword{datasets}
