% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessAlign.R
\name{removeOverlaps}
\alias{removeOverlaps}
\title{Removes overlapping regimens from alignment output}
\usage{
removeOverlaps(df)
}
\arguments{
\item{df}{A dataframe created by align() and processed by prepareDF()}
}
\value{
A data.table object with removed overlaps
}
\description{
Three overlap removal steps are performed:
#' 1. Remove overlaps between regimens with different names  
#' 2. Remove overlaps between regimens with different number of components
#' 3. Remove overlaps between regimens with the same name, same total alignment length, but different scores
}
