% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApps.r
\name{runApps}
\alias{runApps}
\title{Query Data ILOSTAT}
\usage{
runApps(example)
}
\description{
Query data from ILOSTAT SDMX API
}
\details{
Helper function to efficiently query data from ILOSTAT SDMX API.
}
\author{
ILO bescond
}
\seealso{
\code{\link{getCodelist}} \code{\link{getDataStructure}} \code{\link{getData}}
}
\keyword{ILO,}
\keyword{R}
\keyword{SDMX,}

