% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentSummaryReport.R
\name{sameConsecutiveRegimens}
\alias{sameConsecutiveRegimens}
\title{Consecutive regimens are the same?
Same consecutive regimens?}
\usage{
sameConsecutiveRegimens(pa)
}
\arguments{
\item{pa}{Processed alignment data with personID and t_start}
}
\description{
This function checks if consecutive regimens are the same for each patient.
}
