% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.eDNA_model}
\alias{summary.eDNA_model}
\title{Summary of an eDNA model}
\usage{
\method{summary}{eDNA_model}(object, probs = c(0.025, 0.5, 0.975), ...)
}
\arguments{
\item{object}{an object of class eDNA_model_*}

\item{probs}{probabilities for the quantiles of the posterior
samples}

\item{...}{currently ignored}
}
\value{
data.frame, with summary of the fixed effects from the
    model
}
\description{
Summary of eDNA model
}
\details{
Summary of eDNA model
}
\author{
Matt Espe
}
