% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.eDNA_p_detect}
\alias{summary.eDNA_p_detect}
\title{Summary method for eDNA p(detect)}
\usage{
\method{summary}{eDNA_p_detect}(object, probs = c(0.025, 0.5, 0.975), ...)
}
\arguments{
\item{object}{an object of class eDNA_p_detect}

\item{probs}{probabilities for summary, passed to \code{quantile}}

\item{...}{ignored}
}
\value{
a data.frame, with summary statistics of the object
}
\description{
Summary method for eDNA p(detect)
}
\details{
Summary method for eDNA p(detect)
}
\author{
Matt Espe
}
