% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.eDNA_predict_lm}
\alias{summary.eDNA_predict_lm}
\title{Summarize predictions}
\usage{
\method{summary}{eDNA_predict_lm}(object, probs = c(0.025, 0.5, 0.975), FUN = quantile, ...)
}
\arguments{
\item{object}{an object of class eDNA_p_detect}

\item{probs}{probabilities for summary, passed to \code{quantile}}

\item{FUN}{a function to apply to the predictions}

\item{...}{additional arguments passed to FUN}
}
\value{
data.frame
}
\description{
Summarizes predictions into a simple data.frame.
}
\author{
Matt Espe
}
