% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.eDNA_simulation}
\alias{summary.eDNA_simulation}
\title{Summary of eDNA simulations}
\usage{
\method{summary}{eDNA_simulation}(object, var = "Cq_star", probs = c(0.025, 0.5, 0.975), ...)
}
\arguments{
\item{object}{an object of class eDNA_simulation_*}

\item{var}{the simulated variable to summarize, either "Cq_star"
or "ln_conc"}

\item{probs}{probabilities for the summary of the posterior
samples}

\item{...}{currently ignored}
}
\value{
data.frame
}
\description{
Summary of eDNA simulations
}
\details{
Summary of eDNA simulations
}
\author{
Matt Espe
}
