% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_cadastre.R
\name{url_cadastre}
\alias{url_cadastre}
\title{Gives the URL of the cadastre datanase corresponding to the city of the address input in geocode_address}
\usage{
url_cadastre(city_code, context)
}
\arguments{
\item{city_code}{numerical vector from geocode result}

\item{context}{character vector from geocode result}
}
\value{
character vector
}
\description{
Gives the URL of the cadastre datanase corresponding to the city of the address input in geocode_address
}
\examples{
\dontrun{
url_cadastre(75107,"75, Paris")
}

}
