% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{validdrugs}
\alias{validdrugs}
\title{A default set of valid drugs used for alignment}
\format{
A data frame with 382 rows and 8 variables:
}
\usage{
validdrugs
}
\description{
A set of default valid drugs, derived from HemOnc - Thoracic Cancer
}
\keyword{datasets}
