% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputOutput.R
\name{writeOutputs}
\alias{writeOutputs}
\title{Filter a stringDF dataframe to contain only valid patients}
\usage{
writeOutputs(output_all, output_processed, output_eras, regStats)
}
\arguments{
\item{output_all}{A dataframe containing raw outputs}

\item{output_processed}{A dataframe containing processed output regimens}

\item{output_eras}{A dataframe containing processed regimen eras}

\item{regStats}{A dataframe containing various summary statistics}
}
\description{
Filter a stringDF dataframe to contain only valid patients
}
