% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessAlign.R
\name{add_cumultive_times_to_df}
\alias{add_cumultive_times_to_df}
\title{Prepare data.frame for postproccessing. 
Adds cumulative times from drugDF to the dataframe,
where each regimen starts and ends}
\usage{
add_cumultive_times_to_df(df, drugDF)
}
\arguments{
\item{df}{A dataframe with drugRec_Start and drugRec_End columns}

\item{drugDF}{A drugDF object generated by createDrugDF()}
}
\value{
A data.table object without removed overlaps
}
\description{
Prepare data.frame for postproccessing. 
Adds cumulative times from drugDF to the dataframe,
where each regimen starts and ends
}
