% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\title{Aligns two sequences together, returning either a global alignment or a set of local alignments, either overlapping
or otherwise}
\usage{
align(s1, s2, g, Tfac, s = NA, local_align, verbose, mem, removeOverlap)
}
\arguments{
\item{s1}{A sequence to be aligned, usually a short sequence defining a chemotherapy regimen}

\item{s2}{A longer sequence to be aligned, usually a patient's converted drug occurrence table}

\item{g}{A gap penalty supplied to the temporal needleman wunsch/smith waterman algorithms}

\item{Tfac}{The time penalty factor. All time penalties are calculated as a percentage of Tfac}

\item{s}{A substituion matrix, either user-defined or derived from defaultSmatrix. Will be auto-generated if left blank.}

\item{local_align}{A variable defining whether or not alignment should be global (0) or local (1)}

\item{verbose}{A variable indicating how verbose the python script should be in reporting results
Verbose = 0 : Print nothing
Verbose = 1 : Print seqs and scores
Verbose = 2 : Report seqs, scores, H and traceMat}

\item{mem}{A number defining how many sequences to hold in memory during local alignment.
Mem = -1 : Script will define memory length according to len(s1)/len(s2)
Mem = 0 : Script will return exactly 1 alignment
Mem = 1 : Script will return 1 alignment and all alignments with the same score
Mem = X : Script will return X alignments and all alignments with equivalent score as the Xth alignment}

\item{removeOverlap}{A variable indicating whether to remove overlaps (1) or leave them in the output data (0)}
}
\value{
dat A dataframe containing information on the resulting alignments
}
\description{
Aligns two sequences together, returning either a global alignment or a set of local alignments, either overlapping
or otherwise
}
\examples{
dat <- decode(s1,s2)
}
