% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentSummaryReport.R
\name{anyDrugMissingInRegimens}
\alias{anyDrugMissingInRegimens}
\title{Any Drug Missing In Regimens?}
\usage{
anyDrugMissingInRegimens(pa, regimens)
}
\arguments{
\item{pa}{Processed alignment data with personID and t_start.}
}
\description{
This function checks for each patient if there is a drug missing in regimens. 
It uses the anyUncoveredDrugs function to check for each patient separately.
}
