% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artemis-package.R
\docType{package}
\name{artemis-package}
\alias{artemis-package}
\alias{artemis}
\title{The 'artemis' package.}
\description{
The artemis package implements a framework for
    simulating data and fitting models for eDNA studies. In
    particular, it implements a Bayesian latent-variable model in
    which the predictors affect a latent variable, eDNA
    concentration. This latent variable is related to an observed
    variable, CQ cycles via a standard curve
    calibration. CQ cycles, are often truncated at a certain limit,
    where more than X cycles is considered a
    non-detection.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
