% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artemis-package.R
\docType{package}
\name{artemis-package}
\alias{artemis-package}
\alias{artemis}
\title{The 'artemis' package.}
\description{
Simulating data and fitting models for eDNA studies
    can be accomplished using the artemis package. In particular,
    artemis implements a Bayesian latent-variable model in which the
    predictors affect a latent variable, eDNA concentration. This
    latent variable is related to an observed variable, CQ cycles
    via a standard curve calibration. CQ cycles, are often
    truncated at a certain limit, where more than X cycles is
    considered a non-detection.

Additional details on the model implemented by the artemis package
can be found in the "Getting Started" vignette.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
