\name{artemis-package}
\alias{artemis-package}
\alias{artemis}
\docType{package}
\title{
\packageTitle{artemis}
}
\description{
\packageDescription{artemis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{artemis}
\packageIndices{artemis}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{artemis}

Maintainer: \packageMaintainer{artemis}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
