% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{as.data.frame.eDNA_simulation}
\alias{as.data.frame.eDNA_simulation}
\title{Methods for eDNA simulations}
\usage{
\method{as.data.frame}{eDNA_simulation}(x, row.names, optional, ...)
}
\arguments{
\item{x}{object of class eDNA_simulation}

\item{row.names}{ignored}

\item{optional}{ignored}

\item{...}{ignored}
}
\value{
data.frame
}
\description{
Methods for eDNA simulations
}
\details{
as.data.frame methods for eDNA simulations. This allows the
conversion of the simulations to a form suitable for additional
operations, e.g. plotting.
}
\author{
Matt Espe
}
