% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-eras.R
\name{calculateEras}
\alias{calculateEras}
\title{Adds first/second/other era data to a processed regimen alignment ouput}
\usage{
calculateEras(pa, discontinuationTime = 120)
}
\arguments{
\item{pa}{A dataframe processed alignments, generated by processAlignments()}

\item{discontinuationTime}{The number of days to use to indicate treatment discontinuation}
}
\value{
A processed alignment dataframe with added era data relating to first/second/other sequencing
}
\description{
Adds first/second/other era data to a processed regimen alignment ouput
}
