% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{combineAndRemoveOverlaps}
\alias{combineAndRemoveOverlaps}
\title{Combines and removes overlapping regimens from an alignment output}
\usage{
combineAndRemoveOverlaps(output, drugRec, drugDF, regimenCombine)
}
\arguments{
\item{output}{An output dataframe created by align()}

\item{drugRec}{A drug record}

\item{drugDF}{A drugDF object generated by createDrugDF()}

\item{regimenCombine}{Allowed days between same regimen before being combined}
}
\description{
Combines and removes overlapping regimens from an alignment output
}
