% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessAlign.R
\name{combineOverlaps}
\alias{combineOverlaps}
\title{Combines overlapping regimens from alignment output}
\usage{
combineOverlaps(df, regimenCombine)
}
\arguments{
\item{regimenCombine}{Allowed days between same regimen before being combined}

\item{output}{An output dataframe created by align()}
}
\description{
Combines overlapping regimens from alignment output
}
