% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessAlign.R
\name{combineOverlaps}
\alias{combineOverlaps}
\title{Combines overlapping regimens from alignment output
We only merge same regimens, next to each other chronologically
We sort chronologically
Create id for consecutive regimens and remove if distance is larger than regimenCombine}
\usage{
combineOverlaps(df, regimenCombine)
}
\arguments{
\item{regimenCombine}{Allowed days between same regimen before being combined}

\item{output}{An output dataframe created by align()}
}
\description{
Combines overlapping regimens from alignment output
We only merge same regimens, next to each other chronologically
We sort chronologically
Create id for consecutive regimens and remove if distance is larger than regimenCombine
}
