% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{combineOverlaps}
\alias{combineOverlaps}
\title{Combines overlapping regimens from alignment output}
\usage{
combineOverlaps(output, regimenCombine)
}
\arguments{
\item{output}{An output dataframe created by align()}

\item{regimenCombine}{Allowed days between same regimen before being combined}
}
\description{
Combines overlapping regimens from alignment output
}
