% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{createDrugDF}
\alias{createDrugDF}
\title{Creates a plottable data frame from a drug record}
\usage{
createDrugDF(drugRec)
}
\arguments{
\item{drugRec}{A drug record}
}
\description{
Creates a plottable data frame from a drug record
}
