% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generateCohortStats}
\alias{generateCohortStats}
\title{A function to conveniently generate several stats relating to the input cohort}
\usage{
generateCohortStats(connectionDetails, cdmSchema, con_df, stringDF)
}
\arguments{
\item{connectionDetails}{A set of DatabaseConnector connectiondetails}

\item{cdmSchema}{A schema containing a valid OMOP CDM}

\item{con_df}{A con_df dataframe}

\item{stringDF}{A stringDF object containing all valid patients (i.e., those who have exposure
to at least one valid drug)}
}
\value{
A dataframe containing various summary statistics regarding the age of treated
and untreated patients
}
\description{
A function to conveniently generate several stats relating to the input cohort
}
