% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generateRawAlignments}
\alias{generateRawAlignments}
\title{Generate a set of processed alignments given a stringDF dataframe}
\usage{
generateRawAlignments(
  stringDF,
  regimens,
  g,
  Tfac,
  s = NA,
  verbose = 0,
  method,
  writeOut = TRUE,
  outputName = "Output",
  scoreFilter = 0.601
)
}
\arguments{
\item{stringDF}{A stringDF dataframe}

\item{regimens}{A regimen dataframe, containing required regimen shortStrings
for testing}

\item{g}{A gap penalty supplied to the temporal needleman wunsch/smith waterman algorithms}

\item{Tfac}{The time penalty factor. All time penalties are calculated as a percentage of Tfac}

\item{s}{A substituion matrix, either user-defined or derived from defaultSmatrix.
Will be auto-generated if left blank.}

\item{verbose}{A variable indicating how verbose the python script should be in reporting results
Verbose = 0 : Print nothing
Verbose = 1 : Print seqs and scores
Verbose = 2 : Report seqs, scores, H and traceMat}

\item{method}{A character string indicating which loss function method to utilise. Please pick one of
PropDiff        - Proportional difference of Tx and Ty
AbsDiff         - Absolute difference of Tx and Ty
Quadratic       - Absolute difference of Tx and Ty to the power 2
PropQuadratic   - Absolute difference of Tx and Ty to the power 2, divided by the max of Tx and Ty
LogCosh         - The natural logarithm of the Cosh of the absolute difference of Tx and Ty}

\item{writeOut}{A variable indicating whether to save the set of drug records}

\item{outputName}{The name for a given written output}

\item{scoreFilter}{A score filter below which alignments will not be returned. It is recommended to consider
the run time implications if this is set to be below ~0.601. A filter of 0 will result in extremely long
processing times.}
}
\value{
A dataframe containing the relevant patients and their drug exposure strings
}
\description{
Generate a set of processed alignments given a stringDF dataframe
}
