% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generateRegimenStats}
\alias{generateRegimenStats}
\title{Generates a data frame containing summary stats from processed regimen data}
\usage{
generateRegimenStats(processedEras)
}
\arguments{
\item{processedEras}{A dataframe processed alignments, generated by processAlignments()}
}
\value{
A data frame containing summary stats from processed regimen data
}
\description{
Generates a data frame containing summary stats from processed regimen data
}
