% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentSummaryReport.R
\name{generateSummaryReport}
\alias{generateSummaryReport}
\title{Generate summary report}
\usage{
generateSummaryReport(pa, ra, regimens)
}
\arguments{
\item{pa}{Processed alignment data with personID and t_start.}
}
\description{
This function generates a summary report for the processed alignment data. 
It checks if regimens start after the first drug, 
if regimens end before the last drug, 
if consecutive regimens are the same, 
and if there are any uncovered drugs for each patient.
}
