% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getRegimens}
\alias{getRegimens}
\title{Create a dataframe containing regimen data}
\usage{
getRegimens(condition)
}
\arguments{
\item{condition}{A toggle for selecting which condition to retrieve regimens for
Selecting "Deduped" returns all regimens regardless of condition, with duplicates removed}
}
\value{
regDF
regDF <- getRegimens("Deduped")
}
\description{
Create a dataframe containing regimen data
}
