% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessAlign.R
\name{lineOfTreatment}
\alias{lineOfTreatment}
\title{Adds line of treatment to a processed regimen alignment ouput}
\usage{
lineOfTreatment(pa, drugDF, discontinuationTime = 120)
}
\arguments{
\item{pa}{A dataframe processed alignments, generated by processAlignments()}

\item{drugDF}{A drugDF object generated by createDrugDF()}

\item{discontinuationTime}{The number of days to use to indicate treatment discontinuation}
}
\value{
A processed alignment dataframe with added era data relating to first/second/other sequencing
}
\description{
Adds line of treatment to a processed regimen alignment ouput
}
