% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lnconc_to_cq}
\alias{lnconc_to_cq}
\title{Convert [eDNA] to Cq}
\usage{
lnconc_to_cq(x, std_curve_alpha, std_curve_beta, censor = TRUE, upper_Cq = 40)
}
\arguments{
\item{x}{log[eDNA] values}

\item{std_curve_alpha}{the alpha (intercept) value for the
standard curve}

\item{std_curve_beta}{the beta (slope) value for the standard
curve}

\item{censor}{logical, whether Cq values larger than the upper_Cq
should be censored. If TRUE, these any values above the
threshold will be replaced with the upper_Cq value.}

\item{upper_Cq}{the max Cq value}
}
\value{
vector of Cq values
}
\description{
Convenience function for converting values
}
\details{
Convenience function for converting values
}
\author{
Matt Espe
}
