% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputOutput.R
\name{loadDrugs}
\alias{loadDrugs}
\title{Load the default valid drugs dataframe}
\usage{
loadDrugs(absolute = NULL)
}
\arguments{
\item{absolute}{Optional. Absolute path to a custom `.rda` file containing the `validdrugs` dataset. If not provided, the function loads from the `ARTEMIS` package.}
}
\description{
Load the default valid drugs dataframe
}
