% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputOutput.R
\name{loadRegimens}
\alias{loadRegimens}
\title{Load regimens for a given condition
Presently, the only condition fully mapped is lungCancer}
\usage{
loadRegimens(condition)
}
\description{
Load regimens for a given condition
Presently, the only condition fully mapped is lungCancer
}
