% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.eDNA_p_detect}
\alias{plot.eDNA_p_detect}
\title{Plot p(detect)}
\usage{
\method{plot}{eDNA_p_detect}(x, y, probs = c(0.025, 0.975), ylim = c(0,
  1), ...)
}
\arguments{
\item{x}{an object of class "eDNA_p_detect", produced by \code{est_p_detect}}

\item{y}{ignored}

\item{probs}{numeric vector of length 2, the lower and upper probabilities for
quantiles displayed}

\item{ylim}{the y limits on the plot}

\item{...}{additional args passed to \code{plot}}
}
\description{
Plot the p(detect)
}
\details{
Plot the p(detect)
}
\author{
Matt Espe
}
