% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.eDNA_p_detect}
\alias{plot.eDNA_p_detect}
\title{Plot p(detect)}
\usage{
\method{plot}{eDNA_p_detect}(
  x,
  y,
  probs = c(0.025, 0.975),
  ylim = c(0, 1),
  point_size = 2.5,
  n_breaks = 5,
  error_width = rel(1.25),
  ...
)
}
\arguments{
\item{x}{object of class eDNA_p_detect}

\item{y}{ignored}

\item{probs}{probabilities for error bars when a model fit is
supplied}

\item{ylim}{limits on the y axis}

\item{point_size}{point size for the estimated p(detect)}

\item{n_breaks}{passed to \code{pretty}}

\item{error_width}{the width of the ends of errorbars}

\item{...}{additional args passed to plot.stanfit}
}
\value{
a ggplot object
}
\description{
Plot the p(detect eDNA)
}
\details{
This allows the visualization of the probability of detection
producted by \code{est_p_detect}. If the object includes an
interval estimate, errorbars are produced with those intervals.
}
\author{
Matt Espe
}
