% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotAlignment}
\alias{plotAlignment}
\title{Plots a full alignment output
For each patient}
\usage{
plotAlignment(pa, known_drugs = NULL)
}
\arguments{
\item{pa}{A patient alignment dataframe created by align()}
}
\value{
plot - A ggplot object
}
\description{
Plots a full alignment output
For each patient
}
