% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotAlignment}
\alias{plotAlignment}
\title{Plots a full alignment output}
\usage{
plotAlignment(pa, regimens, collapse_regimens = TRUE, add_summary = TRUE)
}
\arguments{
\item{pa}{A patient alignment dataframe created by processAlignments() or generateRawAlignments()}

\item{regimens}{A regimen dataframe, containing required regimen column shortStrings. Used to determine which drugs are missing in regimens for summary report and plot.}

\item{collapse_regimens}{A boolean indicating whether to collapse regimens into a single row}
}
\value{
plot - A list of ggplot objects
}
\description{
For each patient separately, plot drug exposures and aligned regimens over time
}
