% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotFrequency}
\alias{plotFrequency}
\title{Plot Regimen Frequency
Plot frequency of the top N most frequent regimens}
\usage{
plotFrequency(pa, top_n = 10)
}
\arguments{
\item{pa}{Patients alignments dataframe created by processAlignments}

\item{top_n}{Top n most frequent regimens. Default is 10}
}
\value{
plot - A ggplot object
}
\description{
Plot Regimen Frequency
Plot frequency of the top N most frequent regimens
}
