% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotFrequency}
\alias{plotFrequency}
\title{Plots a plot displaying the frequency of the top N most frequent regimens}
\usage{
plotFrequency(processedAll, N = 10)
}
\arguments{
\item{processedAll}{An output dataframe created by processAlignments}

\item{N}{The number of top rows to plot}
}
\description{
Plots a plot displaying the frequency of the top N most frequent regimens
}
