% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotOutput}
\alias{plotOutput}
\title{Plots, or returns a plot, displaying a full alignment output}
\usage{
plotOutput(output, allowOverlaps = F, fontSize = 2, regimenCombine = 1)
}
\arguments{
\item{output}{An output dataframe created by align()}

\item{allowOverlaps}{A boolean indicating whether to leave or remove overlapping regimens}

\item{fontSize}{The desired font size of the text}

\item{regimenCombine}{Allowed number of days between two instances of the same regimen before
those two instances are combined}
}
\value{
regPlot - A ggplot object
}
\description{
Plots, or returns a plot, displaying a full alignment output
}
\examples{
plotOutput(output,F,1)
outputPlot <- plotOutput(output,F,1)
}
