% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotOutput}
\alias{plotOutput}
\title{Plots, or returns a plot, displaying a full alignment output}
\usage{
plotOutput(
  output,
  fontSize = 2.5,
  regimenCombine = 28,
  returnDat = F,
  returnDrugs = F
)
}
\arguments{
\item{output}{An output dataframe created by align()}

\item{fontSize}{The desired font size of the text}

\item{regimenCombine}{Allowed number of days between two instances of the same regimen before}

\item{returnDat}{A toggle to also return a processed data object}

\item{returnDrugs}{A toggle to indicate whether drugs should be returned by processing,
in addition to regimens}
}
\value{
regPlot - A ggplot object
}
\description{
Plots, or returns a plot, displaying a full alignment output
}
