% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRegimenLengthDistribution}
\alias{plotRegimenLengthDistribution}
\title{Plots a plot displaying the observed regimen length distribution for a given regimen, or two given regimens
processed output}
\usage{
plotRegimenLengthDistribution(pa, components = NULL, top_n = 6)
}
\arguments{
\item{pa}{Patients alignments dataframe created by processAlignments}

\item{components}{A set of regimens of interest}

\item{top_n}{Top n most frequent regimens to plot. Ignored if components provided. Default is 6}
}
\value{
plot - A ggplot object
}
\description{
Plots a plot displaying the observed regimen length distribution for a given regimen, or two given regimens
processed output
}
