% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSankey}
\alias{plotSankey}
\title{Plots a sankey diagram displaying the flow between first, second and third regimen eras}
\usage{
plotSankey(processedEras, regGroups, saveLocation = NA, fileName = "Network")
}
\arguments{
\item{processedEras}{An output dataframe created by calculateEras}

\item{regGroups}{A dataframe indicating how to group regimens}

\item{saveLocation}{A file directory location where files may be saved}

\item{fileName}{A filename prefix for saved files}
}
\description{
Plots a sankey diagram displaying the flow between first, second and third regimen eras
}
