% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotScoreDistribution}
\alias{plotScoreDistribution}
\title{Adjusted Score distribution plot}
\usage{
plotScoreDistribution(pa, components = NULL, top_n = 6)
}
\arguments{
\item{pa}{Patients alignments dataframe created by raw or processAlignments}

\item{components}{A set of regimens of interest}

\item{top_n}{Top n most frequent regimens to plot. Ignored if components provided. Default is 6}
}
\value{
plot - A ggplot object
}
\description{
Plot histogram and density of adjusted scores for regimens, or top regimens by frequency
processed output
}
