% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_feuille.R
\name{point_feuille}
\alias{point_feuille}
\title{Produces the reference of the feuille corresponding to the point}
\usage{
point_feuille(banRmatrix, cadastre_sheet_feuilles)
}
\arguments{
\item{banRmatrix}{matrix given out by banR}

\item{cadastre_sheet_feuilles}{corresponding cadaster_sheet in dataframe form}
}
\value{
integer
}
\description{
Produces the reference of the feuille corresponding to the point
}
\examples{
\dontrun{
point_feuille(geocode_address("youradrress"),feuille_cadaster_sheet.json)}

}
