% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessAlign.R
\name{postprocessSinglePatientDF}
\alias{postprocessSinglePatientDF}
\title{Postprocess alignment output}
\usage{
postprocessSinglePatientDF(
  output,
  regimenCombine = 28,
  discontinuationTime = regimenCombine
)
}
\arguments{
\item{output}{An output dataframe created by align()}

\item{regimenCombine}{Allowed number of days between two instances of the same regimen before}
}
\value{
df - A data.frame object
}
\description{
Postprocess alignment output
}
