% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{prepareDF}
\alias{prepareDF}
\title{Prepare data.frame for postproccessing.}
\usage{
prepareDF(output, drugDF)
}
\arguments{
\item{output}{An output dataframe created by align()}

\item{drugDF}{A drugDF object generated by createDrugDF()}
}
\description{
Prepare data.frame for postproccessing.
}
