% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.eDNA_model}
\alias{print.eDNA_model}
\alias{show,eDNA_model-method}
\alias{eDNA_model,}
\alias{ANY-method}
\title{Print eDNA model results}
\usage{
\method{print}{eDNA_model}(x, digits = getOption("digits"), ...)

\S4method{show}{eDNA_model}(object)
}
\arguments{
\item{x}{object of class "eDNA_model_*"}

\item{digits}{number of digits to show}

\item{...}{additional arguments passed to \code{print}}

\item{object}{an object of class eDNA_model}
}
\value{
x
}
\description{
Print eDNA model results
}
\details{
Print method for eDNA model results. By default, this calls
\code{summary} on the estimated effects, i.e. the "betas", and
displays them with additional information about the model fit.
}
\author{
Matt Espe
}
