% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.eDNA_p_detect}
\alias{print.eDNA_p_detect}
\title{Print eDNA p(detect)}
\usage{
\method{print}{eDNA_p_detect}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{object of class "eDNA_p_detect"}

\item{digits}{number of digits to show}

\item{...}{additional arguments passed to \code{print}}
}
\value{
x
}
\description{
Print method for p(detect)
}
\details{
Print method for the results of \code{est_p_detect}. If
\code{est_p_detect} was provided with a single set of beta values,
only these point estimates are printed. If a fit model was
provided, posterior samples of the p(detect) are produced. These
are summarized and printed together with posterior intervals, as
estimated by the quantile method.
}
\author{
Matt Espe
}
