% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.eDNA_simulation}
\alias{print.eDNA_simulation}
\title{Print eDNA simulation}
\usage{
\method{print}{eDNA_simulation}(
  x,
  FUN = summary,
  digits = getOption("digits"),
  show_variables = FALSE,
  ...
)
}
\arguments{
\item{x}{object of class eDNA_simulation}

\item{FUN}{a function to use to summarize the results, default is
\code{summary}}

\item{digits}{number of digits to show}

\item{show_variables}{logical, should the variable levels used in
the simulation be displayed. This is a bit redundent, since
the variable levels are also displayed in the summary, but
this can be helpful to check the results.}

\item{...}{additional arguments passed to \code{print}}
}
\value{
x
}
\description{
Print eDNA simulation
}
\details{
Print method for the results of simulations of an eDNA
experiment. By default, the results are summarized by each
variable level using \code{summary}.
}
\author{
Matt Espe
}
