% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{processAlignments}
\alias{processAlignments}
\title{Perform post-processing on a data frame of raw alignment results}
\usage{
processAlignments(
  rawOutput,
  regimenCombine,
  writeOut = TRUE,
  outputName = "Output_Processed"
)
}
\arguments{
\item{rawOutput}{An output dataframe produced by generateRawAlignments()}

\item{regimenCombine}{The numeric value of days allowed between regimens of the same
name before they are collapsed/summarised into a single regimen}

\item{writeOut}{A variable indicating whether to save the set of drug records}

\item{outputName}{The name for a given written output}
}
\value{
A dataframe containing the relevant patients and their drug exposure strings
}
\description{
Perform post-processing on a data frame of raw alignment results
}
