% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runArtemis.R
\name{runArtemis}
\alias{runArtemis}
\title{Run the ARTEMIS package}
\usage{
runArtemis(
  cdm,
  outputFolder = "Results",
  runMM = FALSE,
  runAML = FALSE,
  runBC = TRUE,
  generateReportOutput = FALSE,
  reportExamples = 5,
  renderReport = FALSE
)
}
\arguments{
\item{cdm}{A CDM reference object created by `CDMConnector::cdmFromCon`}

\item{outputFolder}{The full path to a folder where the results should be saved}

\item{runMM}{Whether to generate and analyse the multiple myeloma cohort}

\item{runAML}{Whether to generate and analyse the acute myeloid leukemia cohort}

\item{runBC}{Whether to generate and analyse the breast cancer cohort}

\item{generateReportOutput}{Whether to generate a Quarto report from saved ARTEMIS outputs}

\item{reportExamples}{Number of example subjects with longer drug records to include per cohort}

\item{renderReport}{Whether to render the Quarto report immediately when Quarto is available}
}
\value{
NULL (invisibly). Results are written to `outputFolder`.
}
\description{
Run the ARTEMIS package
}
