% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runArtemis.R
\name{runArtemis}
\alias{runArtemis}
\title{Run the ARTEMIS pacakge}
\usage{
runArtemis(cdm, outputFolder = "Results")
}
\arguments{
\item{cdm}{A CDM reference object created by `CDMConnector::cdmFromCon`}

\item{outputFolder}{The full path to a folder where the results should be saved}
}
\value{
NULL (invisibly). Results are written to `outputFolder`.
}
\description{
Run the ARTEMIS pacakge
}
