% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputOutput.R
\name{stringDF_from_cdm}
\alias{stringDF_from_cdm}
\title{Generate a set of patient drug record strings from a valid CDM connection and
a valid cohort JSON.}
\usage{
stringDF_from_cdm(cdm, json, name, writeOut = TRUE, outputName = "Output")
}
\arguments{
\item{cdm}{A valid CDM connection object}

\item{json}{A valid Cohort specification JSON}

\item{name}{A name for the given cohort}

\item{writeOut}{A variable indicating whether to save the set of drug records}

\item{outputName}{The name for a given written output
as a local file}
}
\value{
A dataframe containing the relevant patients and their drug exposure strings
}
\description{
Generate a set of patient drug record strings from a valid CDM connection and
a valid cohort JSON.
}
