% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputOutput.R
\name{writeOutputs}
\alias{writeOutputs}
\title{Filter a stringDF dataframe to contain only valid patients}
\usage{
writeOutputs(
  output_all,
  output_processed,
  output_eras,
  regGroups,
  regStats,
  cdm,
  con_df,
  stringDF
)
}
\arguments{
\item{output_all}{A dataframe containing raw outputs}

\item{output_processed}{A dataframe containing processed output regimens}

\item{output_eras}{A dataframe containing processed regimen eras}

\item{regGroups}{The desired regimen grouping variables for Sankey construction}

\item{regStats}{A dataframe containing various summary statistics}

\item{cdm}{The CDM object used to generate the input data to ARTEMIS}

\item{con_df}{The con dataframe object generated by getCohortSet}

\item{stringDF}{A stringDF object containing all valid patients (i.e., those who have exposure
to at least one valid drug)}
}
\description{
Filter a stringDF dataframe to contain only valid patients
}
